MWA-48: refactor: remove unused EventVenue and Services components; add skeleton components for loading states
