+++ src/components/skeleton/EventVenueSkeleton.tsx
@@ -1,7 +1,7 @@
-import { Skeleton } from "@/components/ui/skeleton";
 import { Card, CardContent, CardHeader } from "@/components/ui/card";
+import { Skeleton } from "@/components/ui/skeleton";
 
-function CarouselSkeleton() {
+function EventVenueSkeleton() {
   return (
     <div className="max-w-7xl mx-auto grid grid-cols-1 sm:grid-cols-2 md:grid-cols-3 gap-8 px-6">
       {Array.from({ length: 3 }).map((_, index) => (
@@ -10,9 +10,10 @@ function CarouselSkeleton() {
             <Skeleton className="h-60 w-full" />
           </CardHeader>
           <CardContent>
-            <Skeleton className="h-7 w-40 rounded-md" />
+            <Skeleton className="h-8 w-40 rounded-md" />
             <Skeleton className="h-20 w-full rounded-md mt-2" />
-            <Skeleton className="h-10 w-full rounded-md mt-2" />
+            <Skeleton className="h-6 w-22 rounded-md mt-2" />
+            <Skeleton className="h-9 w-30 rounded-md mt-2" />
           </CardContent>
         </Card>
       ))}
@@ -20,4 +21,4 @@ function CarouselSkeleton() {
   );
 }
 
-export default CarouselSkeleton;
+export default EventVenueSkeleton;
